requirements.txt
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.3.3
requests==2.31.0
python-dotenv==1.0.0
lorem==0.1.1